backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
alembic==1.12.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic[email]==2.4.2
email-validator==2.1.0.post1
python-slugify==8.0.1
httpx==0.25.1
pydantic-settings==2.0.3
python-dateutil==2.8.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
